Remove and ignore db.sqlite3 from git
The dababase file contains data for the running instance of the web app,
such as passwords and application state.  Typically it is not included
in the git repository because it will change whenever one uses the
webapp from git. When getting started on a fresh git checkout, one
should run `python manage.py migrate` to create the database file.