Ignore byte-compiled Python modules
cPython runs in two steps, first it reads in the `*.py` file, parses it,
then outputs a special intermediate representation called Python
byte-code that includes instructions closer to the computer one is
running cPython on.  In most cases cPython will save the byte-code
representation to a file with a file extension of `*.pyc` in a
`__pycache__` directory.  These files change when the version of cPython
changes, or the computer architecture is different (for example: I'm
running on i686 but you're likely on x86_64).

More information here:
https://coderwall.com/p/wrxwog/why-not-to-commit-pyc-files-into-git-and-how-to-fix-if-you-already-did